A linter with an opinion, and a runner that checks every push
ruff now has a configuration instead of its bare defaults: imports, modern syntax, bugbear
and the ruff rules, at the width the code is actually written to. The Russian text in the
code trips the ambiguous-unicode rules, so those three stay off. Everything the new rules
found is fixed — sorted imports, a lambda that captured the loop variable it was called
with, two class-level tables that wanted ClassVar, `int(round(…))`, printf formatting.

The workflow does what a runner can do without a desktop: lint, compile, parse the shell
scripts and load the defaults, checking that config.example.toml has no section the code
does not read.

The `jarvis` alias is gone with it: the command has been `justday` for a while, and
install.sh stopped making the symlink.

+++ src/justday/studio.py
@@ -405,7 +405,7 @@ def _check_locked(job: dict) -> dict:
         try:
             finish(job, src)
             job["state"] = "done"
-        except Exception as e:  # noqa: BLE001 — reported to the user as text
+        except Exception as e:
             job.update(state="failed", error=str(e)[:300])
         if settings()["free_after"]:
             free(url)
@@ -673,7 +673,7 @@ def speech(text: str, out: str | None = None, voice: str = "") -> dict:
 
 
 def _srt_time(t: float) -> str:
-    ms = int(round(t * 1000))
+    ms = round(t * 1000)
     return f"{ms // 3600000:02}:{ms // 60000 % 60:02}:{ms // 1000 % 60:02},{ms % 1000:03}"
